[move] #2 - CustomListTableViewCell 코드 새 파일로 분할. SessionList 폴더 생성
